refac: Refac walkthrough page
