+++ college_flow_app/lib/config/routes/flow_router.dart
@@ -1,7 +1,7 @@
 import 'package:college_flow_app/config/routes/flow_routes.dart';
 import 'package:college_flow_app/presentation/first_screen/first_screen_page.dart';
 import 'package:college_flow_app/presentation/second_screen/second_screen_page.dart';
-import 'package:college_flow_app/presentation/walkthrough_screen/walkthrough_screen.dart';
+import 'package:college_flow_app/presentation/walkthrough/walkthrough_page.dart';
 import 'package:college_flow_app/utils/helpers/navigator_helper.dart';
 import 'package:flutter/material.dart' hide Router;
 
@@ -15,7 +15,7 @@ abstract class FlowRouter {
         break;
       case FlowRoutes.mainPage:
       default:
-        page = const WalkthroughScreen();
+        page = const WalkthroughPage();
         break;
     }
